Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -17,10 +17,10 @@ int main(void)
 
 	float f;
 
-	printf("size of char: %ld byte(s)\n", sizeof(d));
-	printf("size of int : %ld byte(s)\n", sizeof(a));
-	printf("size of a long int: %ld byte(s)\n", sizeof(c));
-	printf("size of long: %ld byte(s)\n", sizeof(b));
-	printf("size of float: %ld byte(s)\n", sizeof(f));
+	printf("size of char: %ld byte(s)\n", (unsigned long)sizeof(d));
+	printf("size of int : %ld byte(s)\n", (unsigned long)sizeof(a));
+	printf("size of a long int: %ld byte(s)\n", (unsigned long)sizeof(c));
+	printf("size of long: %ld byte(s)\n", (unsigned long)sizeof(b));
+	printf("size of float: %ld byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
